Refactor ToolShedRepositoryCache into galaxy.tool_shed.cache
Things required to perform tool shed installs should be galaxy.tool_shed - this module doesn't require anything in galaxy-app - galaxy-data is sufficient.
