+++ mypy.ini
@@ -454,6 +454,8 @@ check_untyped_defs = False
 check_untyped_defs = False
 [mypy-galaxy.tools.cache]
 check_untyped_defs = False
+[mypy-galaxy.tool_shed.cache]
+check_untyped_defs = False
 [mypy-galaxy.tool_util.deps.containers]
 check_untyped_defs = False
 [mypy-galaxy.managers.users]
