+++ test/unit/app/tools/conftest.py
@@ -5,8 +5,8 @@ import pytest
 
 from galaxy.model import tool_shed_install
 from galaxy.model.tool_shed_install import mapping
+from galaxy.tool_shed.cache import ToolShedRepositoryCache
 from galaxy.tool_util.toolbox.base import ToolConfRepository
-from galaxy.tools.cache import ToolShedRepositoryCache
 
 
 @pytest.fixture
